+++ Samples~/Sounds/SoundFx/Scripts/SFXExtensions.cs
@@ -1,5 +1,4 @@
 ﻿using UnityEngine;
-using UnityEngine.Assertions;
 
 namespace Audio.SoundFX
 {
@@ -7,20 +6,12 @@ namespace Audio.SoundFX
     {
         public static void PlaySoundAtLocation(this SFX sfx, Vector3 worldPosition)
         {
-            var sfxManager = SfxManager;
-            
-            Assert.IsNotNull(sfxManager, $"Missing the {nameof(SfxManager)} in the Scene!!");
-            sfxManager.PlaySoundAtLocation(sfx, worldPosition);
+            SFXManager.PlaySoundAtLocation(sfx, worldPosition);
         }
         
         public static void PlaySound(this SFX sfx, float volume = 1f)
         {
-            var sfxManager = SfxManager;
-            
-            Assert.IsNotNull(sfxManager, $"Missing the {nameof(SfxManager)} in the Scene!!");
-            sfxManager.PlaySound(sfx, volume);
+            SFXManager.PlaySound(sfx, volume);
         }
-
-        private static SFXManager SfxManager => SFXManager.Instance;
     }
 }